IoTCloudServices/message_router/code/requirements.txt
paho-mqtt
requests
Flask
Flask-Cors